configuration: Move `ConfigureOptions` to its own file.
