+++ crates/configuration/src/version3/comparison.rs
